Add files via upload
Created the required classes for registration user story. Classes are not complete, missing the database connection for reading, writing or searching for specific data.
for simplicity i used the String data type for storing some values such as phone number, card number and date of reservation because the values are to big for integer or float data type.