Add files via upload
Hello. I created a manager report class for the user story 4. it prompts the manager to input a date which will return the weekly revenue for the week of the specified date, the #of guests, as well as the # of booked rooms for that week. It will read into the reservation and payments java classes and txt files to gather the information. I need to tweak it a bit on my end because its having trouble displaying the messages but i believe its because i haven't configured my text files to the format needed. will fix and update